+++ src/mod-compatibility/all.ts
@@ -1,0 +1,2 @@
+import './nax-art-switch'
+import './lqm-joern-mod'